+++ sdk/highlight-py/CHANGELOG.md
@@ -50,3 +50,10 @@
 ### Fix
 
 - Support `environment` when initializing the SDK.
+
+## v0.6.8 (2023-12-13)
+
+### Fix
+
+- Support tracing auto-instrumentation for the requests library.
+- Add `tracing_origins` configuration to pass X-Highlight-Request header to outgoing requests.